Option for rebuilddb2 is now -T/--ticketspends-batch.
The default is to continuously update the ticket spend info during rebuild, unlike the default for address spend info, since it is very quick with the unspent ticket cache.  The batch update is actually quite slow.

Add the rebuilddb2 option -r/--recoverfromdups, which removes duplicate entries in all tables.  Normally only duplicates in vins and vouts table are removed prior to indexing as these are expected when votes invalidate a block.  Duplicates in the other tables are not expected.